requirements.txt
Flask>=3.0.2
python-dotenv>=1.0.1
waitress>=2.1.2
marshmallow>=3.20.2
pytest>=8.0.0